# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI API Gateway",
    "description": "Open-source AI API Gateway with multi-provider routing and OWASP LLM security guardrails.\n\n## Authentication\nAll endpoints except `GET /gateway/health` and `GET /metrics` require an API key.\nSupply it via **either** header \u2014 the gateway accepts both:\n- `Authorization: Bearer <key>`\n- `X-API-Key: <key>`\n\nWhen `GATEWAY_API_KEYS` is empty (default dev mode) any non-empty key is accepted.\n\n## Security Pipeline (per request)\n1. **Policy enforcement** \u2014 RBAC keyword/topic blocks (`config/policies.yaml`)\n2. **Prompt injection detection** \u2014 heuristic pattern matching\n3. **Secret detection** \u2014 Gitleaks sidecar scans for credentials/tokens\n4. **PII input detection** \u2014 Microsoft Presidio redacts personal data in prompts\n5. **PII output scanning** \u2014 Presidio scans the LLM response before it is returned\n\nRequests with aggregate risk score \u2265 75 are blocked with HTTP 400.\nRisk score and severity are always returned in response headers:\n- `x-gateway-risk-score`\n- `x-gateway-risk-severity`\n- `x-gateway-request-id`\n",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local / Docker Compose"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyHeader": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authorization: Bearer <your-gateway-api-key>"
      },
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "X-API-Key: <your-gateway-api-key>"
      }
    },
    "schemas": {
...
      "StatsResponse": {
        "type": "object",
        "properties": {
          "requests_total": {
            "type": "integer",
            "example": 142
          },
          "requests_blocked": {
            "type": "integer",
            "example": 3
          },
          "pii_detections": {
            "type": "integer",
            "example": 7
          },
          "injection_detections": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "AuditRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "request_id": {
            "type": "string"
          },
          "api_key_hash": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "blocked": {
            "type": "boolean"
          },
          "risk_score": {
            "type": "integer"
          },
          "severity": {
            "type": "string"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "latency_ms": {
            "type": "number"
          },
          "input_tokens": {
            "type": "integer"
          },
          "output_tokens": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuditLogResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditRow"
            }
          }
        }
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Missing API key.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string",
                  "example": "API key required"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Invalid API key.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string",
                  "example": "Invalid API key"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/gateway/stats": {
      "get": {
        "operationId": "gatewayStats",
        "summary": "Gateway Statistics",
        "description": "In-memory counters since last restart (not persisted across restarts).",
        "tags": [
          "Gateway"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current counters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                },
                "example": {
                  "requests_total": 142,
                  "requests_blocked": 3,
                  "pii_detections": 7,
                  "injection_detections": 2
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/gateway/audit": {
      "get": {
        "operationId": "gatewayAudit",
        "summary": "Audit Log",
        "description": "Paginated audit log from SQLite. Contains every request (blocked and passed) with\nrisk scores, provider used, token counts, and latency. API keys are stored as\nSHA-256 hashes only.\n",
        "tags": [
          "Gateway"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of rows to return (1\u20131000).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Rows to skip (for pagination).",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated audit entries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogResponse"
                },
                "example": {
                  "total": 2,
                  "offset": 0,
                  "limit": 100,
                  "data": [
                    {
                      "id": 1,
                      "request_id": "550e8400-e29b-41d4-a716-446655440000",
                      "api_key_hash": "a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3",
                      "model": "gpt-4o-mini",
                      "provider": "openai",
                      "blocked": false,
                      "risk_score": 10,
                      "severity": "low",
                      "reasons": [],
                      "latency_ms": 312.4,
                      "input_tokens": 14,
                      "output_tokens": 9,
                      "created_at": "2024-06-10T12:00:00Z"
                    },
                    {
                      "id": 2,
                      "request_id": "660e8400-e29b-41d4-a716-446655440001",
                      "api_key_hash": "a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3",
                      "model": "gpt-4o-mini",
                      "provider": "\u2014",
                      "blocked": true,
                      "risk_score": 100,
                      "severity": "critical",
                      "reasons": [
                        "Prompt injection pattern matched"
                      ],
                      "latency_ms": 4.1,
                      "input_tokens": 0,
                      "output_tokens": 0,
                      "created_at": "2024-06-10T12:01:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
...
}